# refarch-integrations/refarch-s3-integration/refarch-s3-integration-rest/refarch-s3-integration-rest-core/src/main/java/de/muenchen/refarch/integration/s3/adapter/in/rest/FolderController.java
package de.muenchen.refarch.integration.s3.adapter.in.rest;

import de.muenchen.refarch.integration.s3.adapter.in.rest.dto.FileSizesInFolderDto;
import de.muenchen.refarch.integration.s3.adapter.in.rest.dto.FilesInFolderDto;
import de.muenchen.refarch.integration.s3.adapter.in.rest.dto.FilesMetadataInFolderDto;
import de.muenchen.refarch.integration.s3.adapter.in.rest.mapper.FileMetadataMapper;
import de.muenchen.refarch.integration.s3.adapter.in.rest.mapper.FileSizesInFolderMapper;
import de.muenchen.refarch.integration.s3.adapter.in.rest.mapper.FilesInFolderMapper;
import de.muenchen.refarch.integration.s3.application.port.in.FolderOperationsInPort;
import de.muenchen.refarch.integration.s3.domain.model.FileSizesInFolder;
import de.muenchen.refarch.integration.s3.domain.model.FilesInFolder;
import de.muenchen.refarch.integration.s3.domain.model.FilesMetadataInFolder;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.constraints.NotEmpty;
import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.server.ResponseStatusException;

@Slf4j
@RestController
@RequiredArgsConstructor
@Tag(name = "FolderAPI", description = "API to interact with folders")
@RequestMapping("/folder")
public class FolderController {

    private final FolderOperationsInPort folderOperations;
    private final FilesInFolderMapper filesInFolderMapper;
    private final FileSizesInFolderMapper fileSizesInFolderMapper;
    private final FileMetadataMapper fileMetadataMapper;

    @DeleteMapping
    @Operation(description = "Deletes the folder specified in the parameter")
    public ResponseEntity<Void> deleteFolder(@RequestParam @NotEmpty final String pathToFolder) {
        try {
            log.info("Received a request for deletion of a certain folder.");
            folderOperations.deleteFolder(pathToFolder);
            return ResponseEntity.noContent().build();
        } catch (final Exception exception) {
            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, exception.getMessage(), exception);
        }
    }

    @GetMapping
    @Operation(description = "Returns all file paths for the folder specified in the parameter")
    public ResponseEntity<FilesInFolderDto> getAllFilesInFolderRecursively(@RequestParam @NotEmpty final String pathToFolder) {
        try {
            log.info("Received a request for getting file paths for a certain folder.");
            final FilesInFolder filesInFolder = folderOperations.getAllFilesInFolderRecursively(pathToFolder);
            final FilesInFolderDto filesInFolderDto = this.filesInFolderMapper.model2Dto(filesInFolder);
            return ResponseEntity.ok(filesInFolderDto);
        } catch (final Exception exception) {
            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, exception.getMessage(), exception);
        }
    }

    @GetMapping("/metadata")
    @Operation(description = "Returns metadata of all files for the folder specified in the parameter")
    public ResponseEntity<FilesMetadataInFolderDto> getMetadataOfAllFilesInFolderRecursively(@RequestParam @NotEmpty final String pathToFolder) {
        try {
            log.info("Received a request for getting metadata of all files for a certain folder.");
            final FilesMetadataInFolder filesMetadataInFolder = folderOperations.getMetadataOfAllFilesInFolderRecursively(pathToFolder);
            final FilesMetadataInFolderDto filesMetadataInFolderDto = this.fileMetadataMapper.model2Dto(filesMetadataInFolder);
            return ResponseEntity.ok(filesMetadataInFolderDto);
        } catch (final Exception exception) {
            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, exception.getMessage(), exception);
        }
    }

    @GetMapping("/size")
    @Operation(description = "Returns all file sizes for the folder specified in the parameter")
    public ResponseEntity<FileSizesInFolderDto> getAllFileSizesInFolderRecursively(@RequestParam @NotEmpty final String pathToFolder) {
        try {
            log.info("Received a request for getting file sizes for a certain folder.");
            final FileSizesInFolder fileSizesInFolder = folderOperations.getAllFileSizesInFolderRecursively(pathToFolder);
            final FileSizesInFolderDto filesizesInFolderDto = this.fileSizesInFolderMapper.model2Dto(fileSizesInFolder);
            return ResponseEntity.ok(filesizesInFolderDto);
        } catch (final Exception exception) {
            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, exception.getMessage(), exception);
        }
    }

}